+++ package.json
@@ -19,11 +19,10 @@
   "dependencies": {
     "@tryforge/forge.db": "github:tryforge/ForgeDB#dev",
     "@tryforge/forge.topgg": "github:tryforge/ForgeTopGG#dev",
-    "@tryforge/forgescript": "github:tryforge/Forgescript#dev",
     "axios": "^1.7.9",
     "discord.js": "^14.16.3",
     "dotenv": "^16.4.7",
-    "forge.quirks": "github:LynnuxDev/forge.quirks",
+    "forge.quirks": "github:LynnuxDev/forge.quirks#dev",
     "fs": "0.0.1-security",
     "js-yaml": "^4.1.0",
     "mysql": "^2.18.1",
@@ -34,6 +33,7 @@
     "typescript": "^5.7.2"
   },
   "devDependencies": {
+    "@tryforge/forgescript": "github:tryforge/forgescript#dev",
     "@types/deasync": "^0.1.5",
     "@types/js-yaml": "^4.0.9",
     "@types/node": "^22.10.2",
