+++ src/commands/dev/test.ts
@@ -1,22 +1,4 @@
-import { Command } from '../../types';
+//@ts-ignore
+//@ts-nocheck
 
-const commands: Command[] = [
-  {
-    name: 'test',
-    description: 'test file for devs.',
-    type: 'messageCreate',
-    sourcecode: 'src/commands/dev/test.ts',
-    module: 'dev',
-    version: 'v1.0.0',
-    code: `
-      $onlyIf[$checkContains[$botOwnerID[true;,];$authorID];]
-
-      $httpAddHeader[Authorization;Bearer ${process.env.CROWDIN_API_TOKEN}]
-      $httpRequest[https://api.crowdin.com/api/v2/projects/717569/languages/en/progress;GET;process]
-
-      $logger[Info;$env[process;data;0;data;translationProgress]]
-    `
-  }
-];
-
-export default commands;
+// This file is used to test user commands
